direct sampling contour
Matlab implementation integrated
