+++ examples/calculate_contours_similar_to_docs.py
@@ -1,7 +1,7 @@
 from viroconcom.params import ConstantParam, FunctionParam
 from viroconcom.distributions import WeibullDistribution, LognormalDistribution, \
     MultivariateDistribution
-from viroconcom.contours import IFormContour, HighestDensityContour
+from viroconcom.contours import IFormContour, HighestDensityContour, MonteCarloContour
 import matplotlib.pyplot as plt
 
 # Define a Weibull distribution representing significant wave height.
@@ -24,46 +24,12 @@ mul_dist = MultivariateDistribution(distributions, dependencies)
 
 #print(mul_dist.draw_multivariate_sample(100))
 import numpy as np
-x, y = mul_dist.draw_multivariate_sample(100)
+x, y = mul_dist.draw_multivariate_sample(100000)
 
-x, y = mul_dist.direct_sampling_contour(x, y, 0.1, 5)
-print(x,y)
-def scatter_hist(x, y, ax, ax_histx, ax_histy):
-    # no labels
-    ax_histx.tick_params(axis="x", labelbottom=False)
-    ax_histy.tick_params(axis="y", labelleft=False)
+x, y = MonteCarloContour.direct_sampling_contour(mul_dist, x, y, 0.01, 0.1)
+#print(x, y)
 
-    # the scatter plot:
-    ax.scatter(x, y)
-
-    # now determine nice limits by hand:
-    binwidth = 0.25
-    xymax = max(np.max(np.abs(x)), np.max(np.abs(y)))
-    lim = (int(xymax/binwidth) + 1) * binwidth
-
-    bins = np.arange(-lim, lim + binwidth, binwidth)
-    ax_histx.hist(x, bins=bins)
-    ax_histy.hist(y, bins=bins, orientation='horizontal')
-# definitions for the axes
-left, width = 0.1, 0.65
-bottom, height = 0.1, 0.65
-spacing = 0.005
-
-
-rect_scatter = [left, bottom, width, height]
-rect_histx = [left, bottom + height + spacing, width, 0.2]
-rect_histy = [left + width + spacing, bottom, 0.2, height]
-
-# start with a square Figure
-fig = plt.figure(figsize=(8, 8))
-
-ax = fig.add_axes(rect_scatter)
-ax_histx = fig.add_axes(rect_histx, sharex=ax)
-ax_histy = fig.add_axes(rect_histy, sharey=ax)
-
-# use the previously defined function
-scatter_hist(x, y, ax, ax_histx, ax_histy)
-
+plt.scatter(x, y)
 plt.show()
