+++ viroconcom/contours.py
@@ -585,6 +585,56 @@ class HighestDensityContour(Contour):
         return summed_fields, last_summed
 
 
+class MonteCarloContour(Contour):
+    def __init__(self, mul_var_distribution, deltas=None):
+        None
+
+    def direct_sampling_contour(self, x, y, probability, d_s_deg):
+        """
+        calculates direct sampling contour
+        for fast compution, the data should be 100000 points or less
+        Parameters
+        ----------
+        x,y : array like
+            sample of data
+        probability : float
+            non exceedance probability of contour
+        d_s_deg : float
+            directional step in degrees
+
+        Returns
+        -------
+        x_con, y_con :
+            contour of sample
+        """
+        dt = d_s_deg * np.pi / 180
+        transposed = np.transpose(np.arange(dt, 2 * np.pi, dt))
+        length_x = len(x)
+        length_t = len(transposed)
+        r = np.zeros(length_t)
+
+        # find radius for each angle
+        i = 0
+        while i < length_t:
+            if length_x >= 10000000:
+                print('angle wrong')
+            z = x * np.cos(transposed[i]) + y * np.sin(transposed[i])
+            r[i] = np.quantile(z, probability)
+            i = i + 1
+
+        # find intersection of lines
+        t = np.array(np.concatenate((transposed, [dt]), axis=0))
+        r = np.array(np.concatenate((r, [r[0]]), axis=0))
+
+        denominator = np.sin(t[2:]) * np.cos(t[1:len(t)-1]) - np.sin(t[1:len(t)-1]) * np.cos(t[2:])
+
+        x_cont = (np.sin(t[2:]) * r[1:len(t)-1] - np.sin(t[1:len(t)-1]) * r[2:]) / denominator
+        y_cont = (-np.cos(t[2:]) * r[1:len(r)-1] + np.cos(t[1:len(t)-1]) * r[2:]) / denominator
+
+        return x_cont, y_cont
+
+
+
 if __name__ == "__main__":
     import doctest
     doctest.testmod()
